Implement an initial set of stress test cases for XT systems.
Contains 4 new stress tests.  Adds a templates.py file to draw from a pool of
templates to quickly create new VMs with xec create-vm-with-template.
Modifies experiments.py and test_cases.py to run the new trigger (stress)
and import the TEST_CASES dictionary, respectively.  This is a initial
set of test cases that can (and should) be expanded upon as new ideas for
stress testing are generated.